# src/swagger.json
{
  "openapi": "3.0.1",
  "info": {
    "version": "1.3.0",
    "title": "Stackoveflow clone",
    "description": "A lite clone of Stackoverflow",
    "contact": {
      "name": "Matthew Akhaze"
    }
  },
  "servers": [
    {
      "url": "http://localhost:9001/v1/",
      "description": "Local server"
    }
  ],
  "paths": {
...
    "/votes/question/{id}": {
      "post": {
        "description": "User can vote a question. Cast a vote by sending 'up' or 'down'. User can also change their votes or send the same value again to undo a vote",
        "summary": "Vote question",
        "tags": ["Vote"],
        "operationId": "VoteQuestion",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            },
            "example": "5d99b8a55fd020127207cdee",
            "description": "id of the question to vote for"
          },
          {
            "name": "Authorization",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": ""
          }
        ],
        "requestBody": {
          "required": true,
          "description": "",
          "schema": {
            "example": {
              "vote": "up"
            },
            "type": "object",
            "properties": {
              "vote": {
                "type": "string",
                "enum": ["up", "down"]
              }
            },
            "required": ["vote"]
          }
        },
        "responses": {
          "200": {
            "description": "",
            "schema": {
              "$ref": "#/components/schemas/VoteResponse"
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "VoteResponse": {
        "title": "VoteResponse",
        "example": {
          "result": "SUCCESS",
          "data": { "message": "Question upvoted successfully" }
        },
        "type": "object",
        "properties": {
          "data": {
            "type": "object",
            "properties": {
              "message": {
                "type": "string"
              }
            }
          },
          "result": {
            "type": "string"
          }
        },
        "required": ["result", "data"]
      },
...
    }
  },
...
}